The scraper filers I used
The scraper files as is from when I did the project. I used the 2016 theme and displayed progress on a frontend page template. May modify later so the project is more presentable as a portfolio item but the main functionality is all there, just add the page template file and relevant lines to a 2016 functions.php and set the website URL and it should start working.